fix: resolve portfolio infinite re-render issue
Move filtering logic to ClientPortfolioPage and remove FilterHandler component

#VERCEL_SKIP
