+++ package.json
@@ -65,7 +65,7 @@
     "@types/node": "^22",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "postcss": "^8",
+    "postcss": "^8.5",
     "tailwindcss": "^3.4.17",
     "typescript": "^5"
   }
